+++ front/src/pages/trade/sideBar/sideBarItem/SideBarItem.tsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import caretIcon from '@src/common/utils/caretIcon';
 import formatNumber from '@src/common/utils/formatNumber';
 import { truncateNumber, truncateUnit } from '@src/common/utils/truncateNumber';
-import { IStockListItem } from '@src/recoil/stockList/index';
+import { IStockListItem } from '@src/types';
 import ToggleFavorite from './ToggleFavorite';
 import './SideBarItem.scss';
