+++ front/src/pages/trade/stockInfo/StockInfo.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { IStockListItem } from '@recoil/stockList/index';
+import { IStockListItem } from '@src/types';
 import formatNumber from '@src/common/utils/formatNumber';
 import { truncateNumber, truncateUnit } from '@src/common/utils/truncateNumber';
 import caretIcon from '@src/common/utils/caretIcon';
