+++ front/src/recoil/stockList/index.ts
@@ -1,6 +1,1 @@
-import atom from './atom';
-import type { IStockChartItem, IStockListItem } from './atom';
-
-export { IStockChartItem, IStockListItem };
-
-export default atom;
+export { default } from './atom';
