+++ front/src/types.ts
@@ -26,6 +26,29 @@ export interface IChartItem {
 	amount: number;
 }
 
+export interface IStockChartItem {
+	chartId: number;
+	type: number;
+	priceBefore: number;
+	priceStart: number;
+	priceEnd: number;
+	priceLow: number;
+	priceHigh: number;
+	volume: number;
+	amount: number;
+}
+
+export interface IStockListItem {
+	stockId: number;
+	code: string;
+	nameKorean: string;
+	nameEnglish: string;
+	price: number;
+	previousClose: number;
+	unit: number;
+	charts: IStockChartItem[];
+}
+
 export interface IDailyLog {
 	_id: string;
 	priceEnd: number;
